+++ src/AdminConsole/Pages/App/Settings/Settings.cshtml
@@ -59,102 +59,7 @@
     </div>
 </panel>
 
-<div asp-if="!Model.PendingDelete" class="panel">
-    <h2>Api Keys</h2>
-    <div class="flex space-x-4">
-        <a
-            id="create-public-key-btn"
-            class="btn-primary"
-            asp-page="/App/Settings/CreatePublicKey"
-            asp-route-app="@Model.ApplicationId">
-            Create Public Key
-        </a>
-        <a
-            id="create-secret-key-btn"
-            class="btn-primary"
-            asp-page="/App/Settings/CreateSecretKey"
-            asp-route-app="@Model.ApplicationId">
-            Create Secret Key
-        </a>
-    </div>
-
-    <div class="panel-content">
-        <div class="table-container">
-            <table class="tbl-default min-w-full">
-                <thead>
-                <tr>
-                    <th>Api Key</th>
-                    <th>Type</th>
-                    <th>Locked</th>
-                    <th>Scopes</th>
-                    <th></th>
-                </tr>
-                </thead>
-                <tbody>
-                @if (!Model.ApiKeys.Any())
-                {
-                    <tr>
-                        <td colspan="5">This application has no API Keys or API Secrets.</td>
-                    </tr>
-                }
-                else
-                {
-                    @foreach (var apiKey in Model.ApiKeys)
-                    {
-                        <tr>
-                            <td>@apiKey.Value</td>
-                            <td>@apiKey.Type</td>
-                            @if (apiKey.IsLocked)
-                            {
-                                <td>Yes (<asp-local-time datetime="apiKey.LastLockedAt!.Value"></asp-local-time>)</td>
-                            }
-                            else
-                            {
-                                <td>No</td>
-                            }
-                            <td>
-                                <ul>
-                                    @foreach(var scope in apiKey.Scopes)
-                                    {
-                                        <li>@scope</li>
-                                    }
-                                </ul>
-                            </td>
-                            <td class="flex space-x-2">
-                                @if (apiKey.IsLocked)
-                                {
-                                    <form method="post" asp-page-handler="UnlockApiKey">
-                                        @Html.Hidden(SettingsModel.SelectedApiKeyIdField, apiKey.Id)
-                                        <button class="btn-warning">
-                                            <component type="typeof(UnlockIcon)" render-mode="Static" param-Class="@("h-6 w-6")"/>
-                                        </button>
-                                    </form>
-
-                                }
-                                else
-                                {
-                                    <form method="post" asp-page-handler="LockApiKey">
-                                        @Html.Hidden(SettingsModel.SelectedApiKeyIdField, apiKey.Id)
-                                        <button class="btn-warning">
-                                            <component type="typeof(LockIcon)" render-mode="Static" param-Class="@("h-6 w-6")"/>
-                                        </button>
-                                    </form>
-                                }
-                                <form method="post" asp-page-handler="DeleteApiKey">
-                                    @Html.Hidden(SettingsModel.SelectedApiKeyIdField, apiKey.Id)
-                                    <button class="btn-danger">
-                                        <component type="typeof(DeleteIcon)" render-mode="Static" param-Class="@("h-6 w-6")"/>
-                                    </button>
-                                </form>
-                            </td>
-                        </tr>
-                    }
-                }
-                </tbody>
-            </table>
-        </div>
-    </div>
-</div>
+<partial name="_ApiKeys" model="Model.ApiKeysModel" />
 
 <panel header="Delete Application">
     @if (Model.PendingDelete)
